BOLT 1: clarify handshake by making components more explicit
This commit clarifies the current draft of the crypto spec in several areas by
making the steps more explicit. In addition a few errors in the prior version
of this draft have been fixed.

Thanks @EthanHeilman for pointing out several ambiguities in the prior draft of
the document.

A future commit will contain the addition of several test vectors to ensure
implementation parity.