BOLT 2,3,5: add extra term to offered HTLC output script.
(This is fixed from previous version, where I changed the wrong script!)

Since we expire HTLCs as soon as we can anyway (to avoid one-sided redemption),
we can simply allow very old unspent HTLCs to go to the counterparty.

This makes the witness script 8 bytes longer, and means we no longer need
to receive and store the signatures for each accepted HTLC output to push it
to the HTLC-timeout transaction where we can use our revocation key.  We simply
wait until we can use the extra-long HTLC timeout condition.

(The alternative is to allow spending the HTLC output directly using the
revocation key, but this increases the witness script size by 45 bytes,
though it does simplify some things.  See Pull Request #105).

This does not update the test vectors, as they are still being worked on.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>