BOLT 0,1,2,7: use txout not channel-id for demuxing.
At cost of a few extra bytes, this avoids the whole "oops, we were on
a chain fork" problem.

Now we move the announcement_signature exchange to at least 6 confirms,
which makes re-xmit tricky; I resolved that by insisting on reconnect that
we send if we haven't received, and reply to the first one.

The term "channel shortid" wasn't used anywhere, so I removed it; it's
now a gossip-only thing anyway.

Closes: #114
Suggested-by: Olaoluwa Osuntokun <laolu32@gmail.com>
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>