Remove version, make payment hash a tagged field, force ignoring bad len fields.
It's hard to see how we'd ever change versions, since such a change
would break every existing reader.  We're better off using a new prefix
if we want that, and otherwise preferring backwards compatibility.

On similar lines, we make the payment hash a tagged field, so it too can
change later.

And finally, insist that readers ignore fields they can't understand
because of length: this saves us running introducing a new letter to
eg. replace `p` or `h`.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>