BOLT09: Further specify how bits are specified.
Switched from an ambiguous representation as _bits_ to a bitmask that
is used to check for presence. This was a misunderstanding that Eclair
and c-lightning had. Eclair was interpreting bits=3 as setting the 3rd
bit, whereas c-lightning was interpreting bits=3 as the bitmask to
check for the bits that may be set to indicate mandatory and optional
support.