BOLT0007: add section on ability of nodes to opt out of dynamic updates
This commit adds a new section to the 7th BOLT document detailing a
mechanism that nodes can use to opt-out of dynamic channel updates.
Doing so is primarily a bandwidth optimization, as the node is able to
only receive announcements for new channels and nodes joining the
network, rather than announcements for each channel update within the
network.