Add signature related data in examples
This would have saved me a lot of time and pain. Sure, I should have realized that the document says something paradoxical (signature of the data part.... but the data part includes the signature..... so I should have guessed it meant "excluding the signature"...)

Explicit mention of the signature part might help people. Might not. But helps to be explicit.