BOLT 11: spell out byte limit on `d` field, and note vagueness of `h`.
Our 10-bit length means up to 1023 5-bit words, or 639 bytes.  But `h`
is horribly underspecified since we don't have a general solution: using
a URL is a privacy problem, for example.  It might make sense for a future
transport standard (eg. lightning over HTML, or lightning-over-NFC), but
it's fairly unusable at the moment.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>