BOLT 7: channel_update timestamp must be a UNIX timestamp.
As agreed at the previous meeting, we use the timestamp to prune,
so it does have to be a valid timestamp.  We also suggest ignoring
if it's in the far future, too.

The minutes suggest we can prune for any reason, and that's really
true anyway; the requirements to keep it around are only SHOULD.

Note that this only applies to channel_update: node_announces
are not pruned (except, perhaps, by implication when all channels
are pruned)

Closes: #302
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>