dual-funding: add note about failing if contrib_count is wrong
since we now require the opener to 'pre-send' the total number
of inputs and outputs that they're going to be including in the
funding tx, if the number of actual inputs and outputs that they
send is less than this, we can fail the channel.

we do this because we base our input selection on the number of
inputs and outputs that they send. having a lower number means
that we might be in violation of the rule so we fail this requirement
also