BOLT07: don't send historical updates for gossip_timestamp_filter
This commit removes the suggestion for nodes to respond with all known
updates that satisfy the timestamps of a gossip_timestamp_filter
message. This behavior is identical to using the `initial_routing_sync`,
which has been deprecated due to the inefficient nature of sending the
entire graph on every connection. Now, the requirements simply say that
only new messages should be sent to peers.

Given the upcoming addition of `extended_gossip_queries`,
implementations should use this feature to locate missing graph
information instead of relying the on the remote peer to dump
everything.