BOLT 7: update interpretation of `complete` within reply_channel_range
The current interpretation seems to have been copied over from
reply_short_channel_ids_end. That message should only be sent once at
the end, so the interpretation there makes sense. This is not the case
for reply_channel_range however, since multiple of these messages can be
sent. When responding to a query_channel_range, if the response spans
multiple messages and the message being sent is not the last one, the
field should be `0`, otherwise `1`.