BOLT 11: add optional vendor field.
It was pointed out to me at thelightningconference.com that it's often
a legal requirement to list the vendor on a receipt.  It also makes
perfect sense.

It can be done in the description field, but that's really supposed to
be a description of the *items*.  Dividing it also lets wallets have
much better UX.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>